requirements.txt
podcastfy==0.4.1
python-decouple==3.8
wetro==0.1.21
fastapi==0.115.12
uvicorn==0.27.1
fastapi-cors==0.0.6
pytest-asyncio==0.26.0
python-multipart==0.0.20
pymongo==4.11.0